adding error codes and parameters to scripts

+++ staging/scripts/restore.sh
@@ -2,5 +2,15 @@
 #2. register snapshot in staging domain
 #3. restore from bucket
 
+if (( $# != 1 ))
+then
+	echo "illegal number of parameters, expected 1, got $#"
+	exit 1
+fi
+
+name=${1}
+
 curl -XDELETE 'https://search-histograph-staging-fsuaepsiqkaydkv2w6bxhxmiji.eu-central-1.es.amazonaws.com/_all'
-curl -XPOST 'https://search-histograph-staging-fsuaepsiqkaydkv2w6bxhxmiji.eu-central-1.es.amazonaws.com/_snapshot/histograph-dump/dump_all_11-24-15/_restore'
+curl -XPOST 'https://search-histograph-staging-fsuaepsiqkaydkv2w6bxhxmiji.eu-central-1.es.amazonaws.com/_snapshot/histograph-dump/${name}/_restore'
+
+exit 0
